12.10 12:25 : back on board

+++ bitboard.go
@@ -8,9 +8,6 @@ import (
 
 type bitBoard uint64
 
-// var wPieces uint64
-// var bPieces uint64
-
 func (b bitBoard) count() int {
 	return bits.OnesCount64(uint64(b))
 }
@@ -19,12 +16,12 @@ func (b *bitBoard) set(pos int) {
 	*b |= bitBoard(uint64(1) << uint(pos))
 }
 
-func (b bitBoard) test(pos uint) bool {
-	return (b & bitBoard(uint64(1)<<pos)) != 0
+func (b bitBoard) test(pos int) bool {
+	return (b & bitBoard(uint64(1)<<uint(pos))) != 0
 }
 
-func (b *bitBoard) clr(pos uint) {
-	*b &= bitBoard(^(uint64(1) << pos))
+func (b *bitBoard) clr(pos int) {
+	*b &= bitBoard(^(uint64(1) << uint(pos)))
 }
 
 func (b *bitBoard) firstOne() int {
@@ -32,7 +29,6 @@ func (b *bitBoard) firstOne() int {
 	if bit == 64 {
 		return 64
 	}
-
 	*b = (*b >> uint(bit+1)) << uint(bit+1)
 	return bit
 }
